pyproject.toml
[build-system]
requires = ["setuptools>=80.9"]
build-backend = "setuptools.build_meta"

[project]
name = "transctl"
version = "0.1.0"
description = "A command-line tool for managing application localization and generating translations using machine translation services."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "click~=8.3.1",
    "colorama~=0.4.6",
    "deepl~=1.27.0",
    "pydantic~=2.12.5",
    "click~=8.3.1",
    "colorama~=0.4.6",
    "bs4~=0.0.2",
    "beautifulsoup4~=4.14.3",
    "tomli~=2.4.0",
    "typing_extensions~=4.15.0",
    "tomli_w~=1.2.0",
    "pytest~=9.0.2",
    "SQLAlchemy~=2.0.46",
    "azure-ai-translation-text~=1.0.1"

]

[project.scripts]
transctl = "src.cli.main:main"

[tool.setuptools]
packages = { find = {} }
